requirements.txt
google-api-python-client
google-auth
openai
google-cloud-bigquery
